Move all NavBar related into own NavBar folder
