+++ src/NavBar/nav-menu-item.scss
@@ -1,4 +1,4 @@
-@import "../../variables";
+@import "../variables";
 
 // This cannot be only .core19-nav-menu-item, because then it would be overriden by the normal .btn style
 .btn.core19-nav-menu-item {
